@@ -1,10 +1,10 @@
-import { BrowserModule } from '@angular/platform-browser';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { HttpModule } from '@angular/http';
-import { RouterModule } from '@angular/router';
+import { BrowserModule } from "@angular/platform-browser";
+import { NgModule } from "@angular/core";
+import { FormsModule } from "@angular/forms";
+import { HttpClientModule } from "@angular/common/http"
+import { RouterModule } from "@angular/router";
 
-import { AppComponent } from './app.component';
+import { AppComponent } from "./app.component";
 import { routes, navigatableComponents } from "./app.routing";
 import { ActivityIndicator } from "./components/activity-indicator.component";
 import { GroceryList } from "./pages/list/grocery-list.component";
@@ -21,7 +21,7 @@ import { ItemStatusPipe  } from "./pages/list/item-status.pipe";
   imports: [
     BrowserModule,
     FormsModule,
-    HttpModule,
+    HttpClientModule,
     RouterModule,
     RouterModule.forRoot(routes)
   ],
