@@ -1,7 +1,7 @@
 import { ChangeDetectorRef, ChangeDetectionStrategy, Component, EventEmitter, Input, OnInit, Output, Pipe, PipeTransform } from "@angular/core";
 
-import { Grocery } from "../../shared/grocery/grocery";
-import { GroceryStore } from "../../shared/grocery/grocery-list.service";
+import { Grocery } from "../../shared/grocery/grocery.model";
+import { GroceryListService } from "../../shared/grocery/grocery-list.service";
 
 @Component({
   selector: "grocery-list",
@@ -14,7 +14,6 @@ import { GroceryStore } from "../../shared/grocery/grocery-list.service";
         <span
           [class.done]="grocery.done && !grocery.deleted">{{ grocery.name }}</span>
         <button
-          *ngIf="!grocery.deleted && !grocery.deleting"
           (click)="delete(grocery)">&times;</button>
         <img
           *ngIf="!grocery.deleted && grocery.deleting"
@@ -29,7 +28,7 @@ export class GroceryList {
   @Input() showDeleted: boolean;
   @Output() loaded = new EventEmitter();
 
-  constructor(public store: GroceryStore) {}
+  constructor(public store: GroceryListService) {}
 
   ngOnInit() {
     this.store.load()
@@ -38,31 +37,45 @@ export class GroceryList {
 
   imageSource(grocery) {
     if (grocery.deleted) {
-      return grocery.done ? "./assets/selected.png" : "./assets/nonselected.png"
+      return "./assets/plus.png";
     }
     return grocery.done ? "./assets/checked.png" : "./assets/unchecked.png";
   }
 
   toggleDone(grocery: Grocery) {
     if (grocery.deleted) {
-      grocery.done = !grocery.done;
-      return;
-    }
-
-    this.store.toggleDoneFlag(grocery)
-      .subscribe(
-        () => {},
-        () => { alert("An error occurred managing your grocery list") }
-      );
+      this.store.unsetDeleteFlag(grocery)
+        .subscribe(
+          () => { },
+          () => {
+            alert("An error occurred managing your grocery list.");
+          }
+        );
+    } else {
+      this.store.toggleDoneFlag(grocery)
+        .subscribe(
+          () => { },
+          () => {
+            alert("An error occurred managing your grocery list.");
+          }
+        );
+      }
   }
 
   delete(grocery: Grocery) {
     grocery.deleting = true;
 
-    this.store.setDeleteFlag(grocery)
-      .subscribe(
-        () => {},
-        () => alert("An error occurred while deleting an item from your list.")
-      );
+    let successHandler = () => {};
+    let errorHandler = () => {
+      alert("An error occurred while deleting an item from your list.");
+    };
+
+    if (grocery.deleted) {
+      this.store.permanentlyDelete(grocery)
+        .subscribe(successHandler, errorHandler);
+    } else {
+      this.store.setDeleteFlag(grocery)
+        .subscribe(successHandler, errorHandler);
+    }
   }
 }
