@@ -1,7 +1,7 @@
 import { Component } from "@angular/core";
 import { Router } from "@angular/router";
 
-import { User } from "../../shared/user/user";
+import { User } from "../../shared/user/user.model";
 import { UserService } from "../../shared/user/user.service"
 
 @Component({
